+++ Sources/OsmEdit/OAOpenStreetMapRemoteUtil.mm
@@ -311,7 +311,7 @@ static const NSString* URL_TO_UPLOAD_GPX = @"https://api.openstreetmap.org/api/0
 
 - (OAEntity *)loadEntity:(OATargetPoint *)targetPoint
 {
-    long objectId = targetPoint.obfId;
+    unsigned long long objectId = targetPoint.obfId;
     BOOL isTransportStop = targetPoint.type == OATargetTransportStop;
     if (isTransportStop)
         objectId = ((OATransportStop *)targetPoint.targetObj).poi.obfId;
